add medication statistics
