+++ src/scripts/main.js
@@ -30,4 +30,6 @@ fetchJSON(`${API_ENDPOINT}/reports/stats/`).then(data => {
   joined_items[0].innerText = data.data.msps;
   joined_items[1].innerText = data.data.doctors;
   joined_items[2].innerText = data.data.declarations;
+  joined_items[3].innerText = data.data.pharmacies;
+  joined_items[4].innerText = data.data.pharmacists;
 });
